move mp3_player.h and .c files, update Makefile -ld flags for build
